backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
dj-database-url==3.0.0
Django==4.2.23
djangorestframework==3.16.1
exceptiongroup==1.3.0
fastapi==0.118.0
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.16.0
httplib2==0.31.0
idna==3.10
lxml==6.0.1
numpy==2.0.2
pandas==2.3.2
praw==7.8.1
prawcore==2.4.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.10
pydantic_core==2.33.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytrends==4.9.2
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.14.0
tzdata==2025.2
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
websocket-client==1.8.0
